Project import zIndices and images bugfix
When importing project, if shape had zIndex, prefer
that over setting new zIndex.

Set banner shape config image to be the same as
shape config image, because they are the same.
But when they were about to be drawn, they
would not match resulting in an incorrect draw
on banner.